Moved resultsinfo route to resultsblueprint
