+++ api/templates/_verticalnav.html
@@ -1,5 +1,5 @@
 <div class="vertical-menu">
   <a href="{{ url_for('quiz.create') }}" title="Create a quiz"><button>Create</button></a>
   <a href="{{ url_for('dash.myquizzes') }}" title="See your quizzes"><button>My quizzes</button></a>
-  <a href="{{ url_for('myresults') }}" title="See your quiz results"><button>My Results</button></a>
+  <a href="{{ url_for('dash.myresults') }}" title="See your quiz results"><button>My Results</button></a>
 </div>
