+++ api/templates/account.html
@@ -12,7 +12,7 @@
       
         <div class="action-block">
           <h2>View Your Recent Results</h2>
-          <a href="{{ url_for('myresults') }}" title="See your quiz results" class="button">My Results</a>
+          <a href="{{ url_for('dash.myresults') }}" title="See your quiz results" class="button">My Results</a>
         </div>
         
         <div class="action-block">
