+++ api/templates/resultinfo.html
@@ -29,7 +29,7 @@
     </div>
     
     <div class="results-actions">
-      <a href="{{ url_for('myresults') }}" class="button return-home">Return</a>
+      <a href="{{ url_for('dash.myresults') }}" class="button return-home">Return</a>
       <a href="{{ url_for('taking.quizinfo', quiz_id=result.quiz_id) }}" class="button retry-quiz">Retry</a>
     </div>
   </div>
